move package models and retrofit under data/network
