+++ app/src/main/java/designstring/androidninja/data/network/retrofit/Retrofit.kt
@@ -1,4 +1,4 @@
-package designstring.androidninja.data.retrofit
+package designstring.androidninja.data.network.retrofit
 
 import retrofit2.Retrofit
 import retrofit2.converter.gson.GsonConverterFactory
